withdraw amount
I added withdraw and am trying to figure out how to get the file to save the withdraw and deposit. It adds or subtracts the value from the original value and not the updated.

+++ BankingApp/Transaction.h
@@ -17,7 +17,7 @@ public:
 
 	Transaction(double a, int t) { tAmt = a, type = t; }
 
-	int gettAmt() {
+	double gettAmt() {
 		return tAmt;
 	}
